Menu - new story, relevant to issues #3209, #3212 (#3243)
* refactoring and adding Menu stories

* adding an explanation
