+++ src/js/components/Menu/stories/Simple.js
@@ -1,0 +1,23 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { Grommet, Box, Menu } from 'grommet';
+import { grommet } from 'grommet/themes';
+
+const SimpleMenu = () => (
+  <Grommet theme={grommet}>
+    <Box align="center" pad="large">
+      <Menu
+        dropProps={{ align: { top: 'bottom', left: 'left' } }}
+        label="actions"
+        items={[
+          { label: 'Launch', onClick: () => {} },
+          { label: 'Abort', onClick: () => {} },
+          { label: 'Disabled', disabled: true },
+        ]}
+      />
+    </Box>
+  </Grommet>
+);
+
+storiesOf('Menu', module).add('Simple', () => <SimpleMenu />);
